% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_ca.R
\name{add_census_ca_base_variables}
\alias{add_census_ca_base_variables}
\title{Generate metadata from Canadian census vectors}
\usage{
add_census_ca_base_variables(meta)
}
\arguments{
\item{meta}{tibble with metadata as for example provided by `meta_for_ca_census_vectors`}
}
\value{
tibble with metadata
}
\description{
\lifecycle{maturing}

Add Population, Dwellings, and Household counts to metadata
}
