% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen.R
\name{aggregate_data_with_meta}
\alias{aggregate_data_with_meta}
\title{Aggregate census data up, assumes data is grouped for aggregation
Uses data from meta to determine how to aggregate up}
\usage{
aggregate_data_with_meta(data, meta)
}
\arguments{
\item{data}{census data as obtained from get_census call}

\item{meta}{list with variables and aggregation infromation as obtained from meta_for_vectors}
}
\description{
Aggregate census data up, assumes data is grouped for aggregation
Uses data from meta to determine how to aggregate up
}
