% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_ca_deprecated.R
\name{get_correspondence_ca_census_for}
\alias{get_correspondence_ca_census_for}
\title{Get StatCan DA or DB level correspondence file}
\usage{
get_correspondence_ca_census_for(years, level, refresh = FALSE)
}
\arguments{
\item{years}{list of census years}

\item{level}{geographic level, DA or DB}

\item{refresh}{reload the correspondence files, default is `FALSE`}
}
\value{
tibble with correspondence table`spanning all years
}
\description{
\lifecycle{deprecated}
Joins the StatCan correspondence files for several census years
}
