% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_statcan_correspondence.R
\name{get_single_correspondence_for}
\alias{get_single_correspondence_for}
\title{Get StatCan DA or DB level correspondence file}
\usage{
get_single_correspondence_for(year, level, refresh = FALSE)
}
\description{
Get StatCan DA or DB level correspondence file
}
