% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_ca.R
\name{get_tongfen_ca_census_ct}
\alias{get_tongfen_ca_census_ct}
\title{Grab variables from several censuses on a common geography. Requires sf package to be available
Will return CT level data}
\usage{
get_tongfen_ca_census_ct(
  regions,
  vectors,
  geo_base = NULL,
  match_common_GeoUID = TRUE,
  area_mismatch_cutoff = 0.1,
  tolerance = 500,
  na.rm = FALSE,
  quiet = FALSE,
  refresh = FALSE
)
}
\arguments{
\item{regions}{census region list, should be inclusive list of GeoUIDs across censuses}

\item{vectors}{List of cancensus vectors, can come from different census years}

\item{geo_base}{optional parameter with allowed values one of the geo_datasets to base the geographic data on,}

\item{match_common_GeoUID}{logical, if `TRUE` (default) assumes CTs with same GeoUID are identical}

\item{area_mismatch_cutoff}{cutoff to disregard tracts that can't be matched within threshold}

\item{tolerance}{tolerance in metres when matching changing CT geographies.}

\item{na.rm}{remove NA values when aggregating up values, default is `TRUE`}

\item{quiet}{suppress download progress output, default is `FALSE`}

\item{refresh}{optional character, refresh data cache for this call
for no geographic data, (default `NA`)
or `NULL` (the default) for not returning geographic data`}
}
\value{
dataframe with census variables on common geography
}
\description{
Grab variables from several censuses on a common geography. Requires sf package to be available
Will return CT level data
}
