% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_statcan_correspondence.R
\name{get_tongfen_census_da}
\alias{get_tongfen_census_da}
\title{Grab variables from several censuses on a common geography. Requires sf package to be avaialbe
Will return CT level data}
\usage{
get_tongfen_census_da(regions, vectors, geo_format = NA)
}
\arguments{
\item{regions}{census region list, should be inclusive list of GeoUIDs across censuses}

\item{vectors}{List of cancensus vectors, can come from different census years}

\item{geo_format}{`NA` to only get the variables or 'sf' to also get geographic data}
}
\description{
Grab variables from several censuses on a common geography. Requires sf package to be avaialbe
Will return CT level data
}
