% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_us.R
\name{get_tongfen_us_census_ct}
\alias{get_tongfen_us_census_ct}
\title{Get US census data for 2000 and 2010 census on common census tract based geography}
\usage{
get_tongfen_us_census_ct(state, variables)
}
\arguments{
\item{state}{US state to get the data for}

\item{variables}{vector with census variable names, possibly named}
}
\value{
sf object with (wide form) census variables with census year as suffix (separated by underdcore "_").
}
\description{
Get US census data for 2000 and 2010 census on common census tract based geography
}
