% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen.R
\name{meta_for_additive_variables}
\alias{meta_for_additive_variables}
\title{Generate tongfen metadata for additive variables}
\usage{
meta_for_additive_variables(dataset, variables)
}
\arguments{
\item{dataset}{identifier for the dataset containing the variable}

\item{variables}{(named) vector with additive variables}
}
\value{
a tibble to be used in tongfen_aggregate
}
\description{
\lifecycle{maturing}

Generates metadata to be used in tongfen_aggregate. Variables need to be additive like counts.
}
\examples{
# Get metadata for additive variable Population for the CA16 and CA06 datasets
\dontrun{
meta <- meta_for_additive_variables(c("CA06","CA16"),"Population")
}
}
