% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_ca_deprecated.R
\name{tongfen_ca_census_ct}
\alias{tongfen_ca_census_ct}
\title{Canadian census CT level tongfen via identifier matching}
\usage{
tongfen_ca_census_ct(
  data1,
  data2,
  data2_sum_vars,
  data2_group_vars = c(),
  na.rm = TRUE
)
}
\arguments{
\item{data1}{cancensus CT level dataset for year1 < year2 to serve as base for common geography}

\item{data2}{cancensus CT level dataset for year2 to be aggregated to common geography}

\item{data2_sum_vars}{vector of variable names to by summed up when aggregating geographies}

\item{data2_group_vars}{optional vector of grouping variables}

\item{na.rm}{optional parameter to remove NA values when summing, default = `TRUE`}
}
\value{
`data2` with the variables in `data2_sum_vars` aggregated to a common geography matching `data1`,
identified by the `GeoUID` of `data1`
}
\description{
\lifecycle{deprecated}

Aggregate variables to common CTs, returns data2 on new tiling matching data1 geography
}
