% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen.R
\docType{data}
\name{vancouver_elections_data_2019}
\alias{vancouver_elections_data_2019}
\title{A dataset with polling station votes data from the 2019 federal election in the Vancouver area}
\description{
A dataset with polling station votes data from the 2019 federal election in the Vancouver area
}
\references{
\url{https://www.elections.ca/content.aspx?section=res&dir=rep/off&document=index&lang=e#43GE}
}
\author{
Elections Canada
}
\keyword{data}
