% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen.R
\name{aggregate_data_with_meta}
\alias{aggregate_data_with_meta}
\title{Aggregate census data up, assumes data is grouped for aggregation
Uses data from meta to determine how to aggregate up}
\usage{
aggregate_data_with_meta(data, meta, geo = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{census data as obtained from get_census call, grouped by TongfenID}

\item{meta}{list with variables and aggregation information as obtained from meta_for_vectors}

\item{geo}{logical, should also aggregate geographic data}

\item{na.rm}{logical, should NA values be ignored or carried through.}
}
\value{
data frame with variables aggregated to new common geography
}
\description{
Aggregate census data up, assumes data is grouped for aggregation
Uses data from meta to determine how to aggregate up
}
