% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_statcan_correspondence.R
\name{get_single_correspondence_for}
\alias{get_single_correspondence_for}
\title{Get StatCan DA or DB level correspondence file}
\usage{
get_single_correspondence_for(year, level = c("DA", "DB"), refresh = FALSE)
}
\arguments{
\item{year}{census year}

\item{level}{geographic level, DA or DB}

\item{refresh}{reload the correspondence files, default is `FALSE`}
}
\value{
tibble with correspondence table`
}
\description{
Get StatCan DA or DB level correspondence file
}
