% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_ca.R
\name{meta_for_ca_census_vectors}
\alias{meta_for_ca_census_vectors}
\title{Generate metadata from Candian census vectors}
\usage{
meta_for_ca_census_vectors(vectors)
}
\arguments{
\item{vectors}{list of variables to query}
}
\value{
tidy dataframe with metadata information for requested variables and additional variables
needed for tongfen operations
}
\description{
\lifecycle{maturing}

Build tibble with information on how to aggregate variables given vectors
Queries list_census_variables to obtain needed information and add in vectors needed for aggregation
}
