% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen.R
\name{tongfen_ct}
\alias{tongfen_ct}
\title{Aggregate variables to common cts, returns data2 on new tiling matching data1 geography}
\usage{
tongfen_ct(data1, data2, data2_sum_vars, data2_group_vars = c(), na.rm = TRUE)
}
\arguments{
\item{data1}{Cancensus CT level datatset for year1 < year2 to serve as base for common geography}

\item{data2}{Cancensus CT level datatset for year2 to be aggregated to common geography}

\item{data2_sum_vars}{vector of variable names to by summed up when aggregating geographies}

\item{data2_group_vars}{optional vector of grouping variables}

\item{na.rm}{optional parameter to remove NA values when summing, default = `TRUE`}
}
\description{
Aggregate variables to common cts, returns data2 on new tiling matching data1 geography
}
