% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen.R
\name{tongfen_estimate}
\alias{tongfen_estimate}
\title{Estimate variable values for custom geography}
\usage{
tongfen_estimate(target, source, meta)
}
\arguments{
\item{target}{custom geography to estimate values for}

\item{source}{input geography with values}

\item{meta}{metadata for variable aggregation}
}
\description{
\lifecycle{maturing}

Estimates data from source geometry onto target geometry
}
