% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen.R
\name{tongfen_estimate}
\alias{tongfen_estimate}
\title{Estimate variable values for custom geography}
\usage{
tongfen_estimate(data1, data2, data2_sum_vars, unique_key = NA)
}
\arguments{
\item{data1}{custom geography to estimate values for}

\item{data2}{input geography with values}

\item{data2_sum_vars}{columns of data2 to be summed up}

\item{unique_key}{optional unique key for each geography in data1}
}
\description{
Estimate variable values for custom geography
}
