% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tongfen_ca.R
\name{tongfen_estimate_ca_census}
\alias{tongfen_estimate_ca_census}
\title{Tongfen estimate data for given geometry}
\usage{
tongfen_estimate_ca_census(geometry, level, vectors, na.rm = FALSE)
}
\arguments{
\item{geometry}{geometry}

\item{level}{level to use for tongfen}

\item{vectors}{census variables to aggregate data for, can be named.}

\item{na.rm}{how to deal with NA values, default is \code{FALSE}.}
}
\description{
\lifecycle{maturing}

Estimates values for the given census vectors for the given geometry using
data from the specified level range
}
